FINAL COMMIT for Assignment 1
Okay, I think it should be working now. Still having issues with
cross-browser testing. The furthest I seem to be able to go back is
IE9. Netrender just returns an error message in German for anything
else, and other testing services require a paid account. Not sure what
the fix would be. Maybe it’s something in my code?